Moved missing controls from popup component to control component library.
