+++ src/components/controls/trashcan.tsx
