+++ src/packages/class-diagram/classifier/classifier-popup.tsx
@@ -3,14 +3,13 @@ import { connect } from 'react-redux';
 import { compose } from 'redux';
 import styled from 'styled-components';
 import { ClassElementType } from '..';
-import { Button } from '../../../components/controls/button/button';
 import { Divider } from '../../../components/controls/divider/divider';
+import { Header } from '../../../components/controls/header';
 import { Switch } from '../../../components/controls/switch/switch';
 import { Textfield } from '../../../components/controls/textfield/textfield';
+import { TrashCanIcon } from '../../../components/controls/trashcan';
 import { I18nContext } from '../../../components/i18n/i18n-context';
 import { localized } from '../../../components/i18n/localized';
-import { Header } from '../../../components/popup/controls/header';
-import { TrashCanIcon } from '../../../components/popup/controls/trashcan';
 import { ModelState } from '../../../components/store/model-state';
 import { Element } from '../../../services/element/element';
 import { ElementRepository } from '../../../services/element/element-repository';
