+++ src/packages/use-case-diagram/use-case-association/use-case-association-popup.tsx
@@ -6,10 +6,10 @@ import { UseCaseRelationshipType } from '..';
 import { Divider } from '../../../components/controls/divider/divider';
 import { Dropdown } from '../../../components/controls/dropdown/dropdown';
 import { FlipIcon } from '../../../components/controls/flip-icon';
+import { Header } from '../../../components/controls/header';
 import { Textfield } from '../../../components/controls/textfield/textfield';
 import { I18nContext } from '../../../components/i18n/i18n-context';
 import { localized } from '../../../components/i18n/localized';
-import { Header } from '../../../components/popup/controls/header';
 import { Element } from '../../../services/element/element';
 import { ElementRepository } from '../../../services/element/element-repository';
 import { RelationshipRepository } from '../../../services/relationship/relationship-repository';
